Move sync-exception fallback test to its own file
bun-build-api.test.ts has several heavy tests that brush the 5s default
timeout under load; putting the regression test alongside
plugin-error-nested-throw.test.ts keeps the gated file small and
deterministic.
