+++ test/bundler/bun-build-api.test.ts
@@ -1,7 +1,7 @@
 import assert from "assert";
 import { afterEach, describe, expect, test } from "bun:test";
 import { readFileSync, writeFileSync } from "fs";
-import { bunEnv, bunExe, isASAN, isDebug, tempDir, tempDirWithFiles, tempDirWithFilesAnon } from "harness";
+import { bunEnv, bunExe, isASAN, isDebug, tempDirWithFiles, tempDirWithFilesAnon } from "harness";
 import path, { join } from "path";
 import { buildNoThrow } from "./buildNoThrow";
 
@@ -1178,85 +1178,3 @@ test.skipIf(!isDebug && !isASAN)(
   },
   120_000,
 );
-
-// Regression: the C++ synchronous-exception fallback in JSBundlerPlugin__matchOnLoad /
-// JSBundlerPlugin__matchOnResolve passed the wrong `which` value to JSBundlerPlugin__addError,
-// so Zig would reinterpret a *Load as a *Resolve (and vice versa) and read garbage / crash.
-// The builtin calls the public `.then` on the pending IIFE promise, so we can force a
-// synchronous throw there to reach that fallback deterministically.
-describe.each(["onLoad", "onResolve"] as const)("Bun.build plugin %s builtin throws synchronously", hook => {
-  test.concurrent("addError receives the correct context type", async () => {
-    const fixture = /* js */ `
-      const originalThen = Promise.prototype.then;
-      let armed = false;
-      Promise.prototype.then = function (...args) {
-        if (armed) {
-          armed = false;
-          Promise.prototype.then = originalThen;
-          throw new Error("synchronous throw from ${hook} builtin");
-        }
-        return originalThen.apply(this, args);
-      };
-
-      const result = await Bun.build({
-        entrypoints: [Bun.fileURLToPath(new URL("./entry.ts", import.meta.url))],
-        throw: false,
-        plugins: [
-          {
-            name: "sync-throw",
-            setup(build) {
-              build.${hook}(
-                { filter: ${hook === "onLoad" ? "/entry\\.ts$/" : "/^sync-throw:thing$/"} },
-                () => {
-                  armed = true;
-                  // Pending promise: the outer async IIFE in the builtin suspends on it,
-                  // so the builtin falls through to the public .then() call which throws
-                  // synchronously and surfaces to the C++ TOP_EXCEPTION_SCOPE fallback.
-                  return new Promise(() => {});
-                },
-              );
-              ${hook === "onResolve" ? 'build.onLoad({ filter: /.*/, namespace: "sync-throw" }, () => ({ contents: "", loader: "js" }));' : ""}
-            },
-          },
-        ],
-      });
-
-      Promise.prototype.then = originalThen;
-      console.log(JSON.stringify({
-        success: result.success,
-        logs: result.logs.map(l => String(l.message ?? l)),
-      }));
-    `;
-
-    using dir = tempDir(`bun-build-plugin-sync-throw-${hook}`, {
-      "entry.ts": hook === "onLoad" ? `export const x = 1;` : `import "sync-throw:thing"; export const x = 1;`,
-      "build.ts": fixture,
-    });
-
-    await using proc = Bun.spawn({
-      cmd: [bunExe(), "run", join(String(dir), "build.ts")],
-      env: bunEnv,
-      cwd: String(dir),
-      stdout: "pipe",
-      stderr: "pipe",
-      timeout: 30_000,
-    });
-
-    const [stdout, stderr, exitCode] = await Promise.all([proc.stdout.text(), proc.stderr.text(), proc.exited]);
-
-    // On the buggy build this reinterprets the Load/Resolve struct, which under ASAN
-    // typically aborts the process; on release it may hang or print garbage.
-    expect({
-      stderr: stderr.trim(),
-      exitCode,
-      signalCode: proc.signalCode ?? null,
-    }).toMatchObject({
-      exitCode: 0,
-      signalCode: null,
-    });
-
-    const parsed = JSON.parse(stdout.trim());
-    expect(parsed.success).toBe(false);
-    expect(parsed.logs.join("\n")).toContain(`synchronous throw from ${hook} builtin`);
-  });
-});
